fbsd/ena: update ENA FreeBSD driver to v2.1.0
Changes since last release (v2.0.0)

**New Features**
* Add netmap support.
* Add LLQ support for the aarch64 by enabling WC.

**Bug Fixes**
* Fix infinite missing keep alive reset loop in case the reset routine
  takes too long because of the host being overloaded.

**Minor Changes**
* Split controller and datapath code into separate files.
