+++ kernel/fbsd/ena/README
@@ -3,7 +3,7 @@ FreeBSD kernel driver for Elastic Network Adapter (ENA) family:
 
 Version:
 ========
-v2.0.0
+v2.1.0
 
 Supported FreeBSD Versions:
 ===========================
@@ -139,6 +139,17 @@ ena_eth_io_defs.h - Definition of ENA data path interface.
 ena_common_defs.h - Common definitions for ena_com layer.
 ena_regs_defs.h   - Definition of ENA PCI memory-mapped (MMIO) registers.
 ena_plat.h        - Platform dependent code for FreeBSD.
+
+Overview:
+=========
+
+The driver is supporting the netmap framework. In order to use it, the Makefile
+of the module should be modified and it needs uncommenting of the line:
+# CFLAGS += -DDEV_NETMAP
+
+The kernel must be also built with DEV_NETMAP option in order to be able to use
+the driver with the netmap support, which is default for amd64 but not for
+aarch64.
 
 Management Interface:
 =====================
